fuck CSS3 flexbox - using js instead
CSS3 flexbox is not stable and there are different versions with
different syntax (& may change in the future).

Different browsers behave differently (e.g. IE has different default
values to some attributes)

So many variables to cater for. So using a js solution instead. If js
not supported or disabled, tough shit