fixed slide resize bug at breakpoint (questionable)
I couldn't get a DOM property event to fire on currentSrc change, so I
have done it in a hacky way with the use of a timer every time the
window is resized.

This is a BAD solution as it assumes the src of the image will change in
x seconds. If it takes longer than that, nothing happens.